# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "سیستم مدیریت نوبت‌دهی کلینیک دندانپزشکی",
    "description": "API کامل و حرفه‌ای برای مدیریت نوبت‌دهی دو کلینیک دندانپزشکی\n\n⚠️ نکته مهم CSRF Protection:\n1. ابتدا از /auth/csrf-token یک token دریافت کنید\n2. در header درخواست‌های POST/PATCH/DELETE، X-CSRF-Token را با مقدار token ارسال کنید",
    "version": "1.4.0",
    "contact": {
      "name": "Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Alternative port"
    }
  ],
...
  "paths": {
...
    "/clinics": {
      "get": {
        "tags": ["Clinics"],
        "summary": "لیست کلینیک‌ها",
        "description": "دریافت لیست کلینیک‌ها با pagination",
        "operationId": "getClinics",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "لیست کلینیک‌ها",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Clinic" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Clinics"],
        "summary": "ایجاد کلینیک (فقط مدیر)",
        "operationId": "createClinic",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "address", "phoneNumber"],
                "properties": {
                  "name": { "type": "string", "example": "کلینیک دندانپزشکی تهران" },
                  "address": { "type": "string", "example": "تهران، خیابان ولیعصر، پلاک 123" },
                  "phoneNumber": { "type": "string", "example": "021-12345678" },
                  "description": { "type": "string", "example": "کلینیک مجهز دندانپزشکی" },
                  "latitude": { "type": "number", "example": 35.6892, "description": "عرض جغرافیایی (X)" },
                  "longitude": { "type": "number", "example": 51.3890, "description": "طول جغرافیایی (Y)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "کلینیک ایجاد شد"
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/clinics/{id}": {
      "get": {
        "tags": ["Clinics"],
        "summary": "دریافت کلینیک",
        "operationId": "getClinic",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": {
            "description": "کلینیک",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Clinic" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Clinics"],
        "summary": "به‌روزرسانی کلینیک (مدیر/منشی)",
        "operationId": "updateClinic",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phoneNumber": { "type": "string" },
                  "description": { "type": "string" },
                  "latitude": { "type": "number", "description": "عرض جغرافیایی (X)" },
                  "longitude": { "type": "number", "description": "طول جغرافیایی (Y)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "کلینیک به‌روزرسانی شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Clinics"],
        "summary": "حذف کلینیک (فقط مدیر)",
        "operationId": "deleteClinic",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "کلینیک حذف شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/contact": {
      "get": {
        "tags": ["Contact"],
        "summary": "لیست پیام‌های تماس (فقط مدیر)",
        "operationId": "getContactMessages",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "read",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] },
            "description": "فیلتر بر اساس وضعیت خوانده شده"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "جستجو در نام، ایمیل، موضوع و متن پیام"
          }
        ],
        "responses": {
          "200": {
            "description": "لیست پیام‌ها",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "messages": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ContactMessage" }
                        }
                      }
                    },
                    "meta": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Contact"],
        "summary": "ارسال پیام تماس با ما (عموم)",
        "operationId": "createContactMessage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "message"],
                "properties": {
                  "name": { "type": "string", "example": "احمد احمدی" },
                  "email": { "type": "string", "format": "email", "example": "ahmad@example.com" },
                  "phoneNumber": { "type": "string", "example": "09123456789" },
                  "subject": { "type": "string", "example": "سوال درباره خدمات" },
                  "message": { "type": "string", "example": "متن پیام..." }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "پیام ارسال شد" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/contact/stats": {
      "get": {
        "tags": ["Contact"],
        "summary": "آمار پیام‌های تماس (فقط مدیر)",
        "operationId": "getContactMessagesStats",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "آمار پیام‌ها",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "unread": { "type": "integer" },
                        "read": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/contact/{id}": {
      "get": {
        "tags": ["Contact"],
        "summary": "دریافت پیام تماس (فقط مدیر)",
        "operationId": "getContactMessage",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": {
            "description": "پیام تماس",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": { "$ref": "#/components/schemas/ContactMessage" }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Contact"],
        "summary": "حذف پیام تماس (فقط مدیر)",
        "operationId": "deleteContactMessage",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "پیام حذف شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clinic": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "phoneNumber": { "type": "string" },
          "description": { "type": "string" },
          "latitude": { "type": "number", "example": 35.6892, "description": "عرض جغرافیایی (X) برای نقشه" },
          "longitude": { "type": "number", "example": 51.3890, "description": "طول جغرافیایی (Y) برای نقشه" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ContactMessage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phoneNumber": { "type": "string" },
          "subject": { "type": "string" },
          "message": { "type": "string" },
          "read": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "شناسه یکتای رکورد"
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "description": "شماره صفحه"
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        },
        "description": "تعداد رکوردها در هر صفحه"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "درخواست نامعتبر",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "دسترسی ممنوع - نیاز به نقش مناسب",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "رکورد یافت نشد",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dental.sid",
        "description": "Session-based authentication using cookies"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token برای حفاظت در برابر حملات CSRF. ابتدا /auth/csrf-token را صدا کنید و token دریافتی را در این header ارسال کنید"
      }
    }
  }
}